Rework timer to work as expected when sleeping
I've run into this problem where the on-screen timer doesn't update the countdown in some circumstances. Usually this is just a mildly annoying/humorous glitch -- the countdown says that there's "39 minutes left" when we all know that's not quite right -- but it came up in both of my recent on-site classes and I wanted to investigate a bit.

The old timer works something like this:

```
duration = x
every second:
  duration = duration - 1
  display duration
```

The problem with this is if something interrupts the `every second` check, the duration doesn't shrink.

The new timer works more like this:

```
time to stop = x
every second:
  duration = how long until time to stop
  display duration
```

Because I'm a nerd, I restructured the JS code a bit, while I was in here. It's arguably unnecessary, but it felt good.