update gems and vendor/cache them
github-pages is bumped to being at version 45 because that's what we do on guides, and consistency is awesome.

updating to version 45 of github-pages made me need to update rdiscount to 2.1.8 because of github-pages dependency.

we need to install pygments because though we're saying we have it for highlighter, we never included it?