Remove strikethrough for strangers
Strikethrough is not currently supported with Minimal Mistakes or Jekyll. We would need to implement something a little heavier, like [our own embedded JS](http://stackoverflow.com/questions/17004095/jekyll-markdown-strikethrough). I'm removing the strikethrough at this time since it isn't vital to the course. We can revisit in the future if this is a desired feature.