Submodule cheatsheet: add 'Pulling Superproject Updates' section
When pulling in a repository that has submodules, the default behaviour of
`git pull` is to fetch the submodule, but not to update the working
directory of the submodule.

This leads to a very confusing situation, since just after the pull
the working directory of the submodule is not in sync with the commit
recorded in the superproject.

To complete the update, a separate `git submodule update` is required,
or `git pull` must be called with the `--recurse-submodules` flag.

Also, if the superproject changes the URL of the submodule,
`git submodule sync` is required to synchronize the URL to the local
.git/config and .git/modules/<name>/config configuration files.