Submodule cheatsheet: add 'Changing branches' section
When changing branches in a project that has submodules, the default
behaviour of `git checkout` is to not update the working
directory of the submodule.

This leads to a very confusing situation, since just after the checkout
the working directory of the submodule is not in sync with the commit
recorded in the superproject.

To complete the update, a separate `git submodule update` is required,
or `git checkout` must be called with the `--recurse-submodules` flag.